# strings: speed up Replace